+++ NEWS.md
@@ -1,4 +1,5 @@
 # elections2 0.6.0
 
 * Added a `NEWS.md` file to track changes to the package.
-* Added ability to select results columns by number to wrangle_results() function.
+* Added ability to select results columns by number in the wrangle_more_cols() function.
+* Added ability to use an existing R data frame in wrangle_results() and wrangle_more_cols(). The data frame name must be in quotation marks.